perf: démarrage non-bloquant + lazy loading composants lourds
- DB: sql.js WASM préchauffé en parallèle avec la fenêtre (prewarmSqlJs)
- DB: lecture async du fichier (fs.promises.readFile)
- DB: save après migrations seulement si des migrations ont été appliquées
- Main: fenêtre créée immédiatement, DB ouverte en parallèle, event db:ready envoyé
- Renderer: attend db:ready avant loadCompetitions (fallback 1s)
- App.tsx: CompetitionView lazy-loadé (React.lazy, 57KB différés)
- CompetitionView: PoolView, TableauView, RemoteScoreManager, KioskDisplay,
  FencerComparison, AnalyticsDashboard, PresentationMode, QuestPhaseView en lazy
- Nouveau fichier tableau/tableauTypes.ts: types et propagateWinners extraits
  de TableauView pour permettre son lazy-loading sans casser les imports
- Suspense wrappers ajoutés autour de chaque composant lazy
