+++ src/renderer/components/tableau/TableauPendingSection.tsx
@@ -5,7 +5,7 @@
  */
 
 import React from 'react';
-import { TableauMatch } from '../TableauView';
+import { TableauMatch } from './tableauTypes';
 
 interface TableauPendingSectionProps {
   round: number;
