+++ src/renderer/components/tableau/TableauScoreModal.tsx
@@ -5,7 +5,7 @@
  */
 
 import React from 'react';
-import { TableauMatch } from '../TableauView';
+import { TableauMatch } from './tableauTypes';
 
 interface TableauScoreModalProps {
   match: TableauMatch;
